[chore] upgrade toolchain to nightly-2023-04-24 (#1373)
### Description

Upgrade toolchain to nightly-2023-04-24 for future compatibilities.

### Issue Link

N/A

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

### Contents

- Toolchain upgrade to nightly-2023-04-24
- Fix lint errors

### Rationale

The toolchain currently in use is outdated and lacks stable features.
Issue #1372 requires GAT, which has been stabilized since version
`1.65.0`. Meanwhile, the current stable Rust version is `1.69.0`.

### How Has This Been Tested?

N/A

+++ circuit-benchmarks/src/pi_circuit.rs
@@ -68,7 +68,7 @@ mod tests {
         // Bench setup generation
         let setup_message = format!("{} {} with degree = {}", BENCHMARK_ID, setup_prfx, degree);
         let start1 = start_timer!(|| setup_message);
-        let general_params = ParamsKZG::<Bn256>::setup(degree as u32, &mut rng);
+        let general_params = ParamsKZG::<Bn256>::setup(degree, &mut rng);
         let verifier_params: ParamsVerifierKZG<Bn256> = general_params.verifier_params().clone();
         end_timer!(start1);
