+++ zkevm-circuits/src/evm_circuit/execution/extcodecopy.rs
@@ -226,7 +226,7 @@ impl<F: Field> ExecutionGadget<F> for ExtcodecopyGadget<F> {
             region,
             offset,
             memory_length.as_u64(),
-            memory_expansion_gas_cost as u64,
+            memory_expansion_gas_cost,
         )?;
 
         Ok(())
