+++ zkevm-circuits/src/evm_circuit/execution/msize.rs
@@ -67,11 +67,8 @@ impl<F: Field> ExecutionGadget<F> for MsizeGadget<F> {
         step: &ExecStep,
     ) -> Result<(), Error> {
         self.same_context.assign_exec_step(region, offset, step)?;
-        self.value.assign(
-            region,
-            offset,
-            Some((step.memory_size as u64).to_le_bytes()),
-        )?;
+        self.value
+            .assign(region, offset, Some((step.memory_size).to_le_bytes()))?;
 
         Ok(())
     }
