+++ zkevm-circuits/src/keccak_circuit/cell_manager.rs
@@ -59,11 +59,7 @@ impl<F: FieldExt> Cell<F> {
     }
 
     pub(crate) fn assign(&self, region: &mut KeccakRegion<F>, offset: i32, value: F) {
-        region.assign(
-            self.column_idx,
-            ((offset as i32) + self.rotation) as usize,
-            value,
-        );
+        region.assign(self.column_idx, (offset + self.rotation) as usize, value);
     }
 
     pub(crate) fn assign_value(&self, region: &mut KeccakRegion<F>, offset: i32, value: Value<F>) {
@@ -73,11 +69,7 @@ impl<F: FieldExt> Cell<F> {
         // this shouldn't be needed.
         let value_f = extract_field(value);
 
-        region.assign(
-            self.column_idx,
-            ((offset as i32) + self.rotation) as usize,
-            value_f,
-        );
+        region.assign(self.column_idx, (offset + self.rotation) as usize, value_f);
     }
 }
